Adding more automated tests

+++ src/FairPlayBudgetSln/FairPlayBudget.Interfaces/Services/IMonthlyBudgetInfoService.cs
@@ -11,12 +11,12 @@ namespace FairPlayBudget.Interfaces.Services
     {
         Task CreateMonthlyBudgetInfoAsync(CreateMonthlyBudgetInfoModel createMonthlyBudgetInfoModel,
             CancellationToken cancellationToken);
-        Task UpdateMonthlyBudgetInfoAsync(int monthlyBudgetInfoModelId, CreateMonthlyBudgetInfoModel createMonthlyBudgetInfoModel,
+        Task UpdateMonthlyBudgetInfoAsync(long monthlyBudgetInfoModelId, CreateMonthlyBudgetInfoModel createMonthlyBudgetInfoModel,
             CancellationToken cancellationToken);
 
         Task<MonthlyBudgetInfoModel[]?> GetMyMonthlyBudgetInfoListAsync(CancellationToken cancellationToken);
         Task<CreateMonthlyBudgetInfoModel> LoadMonthlyBudgetInfoAsync(
-            int monthlyBudgetInfoId,
+            long monthlyBudgetInfoId,
             CancellationToken cancellationToken);
         Task<CreateMonthlyBudgetInfoModel> ImportFromTransactionsFileStreamAsync(Stream stream, CancellationToken cancellationToken);
         Task<CreateMonthlyBudgetInfoModel> ImportFromCreditCardFileStreamAsync(Stream stream, CancellationToken cancellationToken);
